requirements.txt
typer[all]>=0.9.0
Pillow>=9.0.0
numpy>=1.21
scikit-learn>=1.0
scikit-image>=0.19
scipy>=1.7
